refactor: migrate pipeline, probe, and price methods to sub-modules — damai_app.py 2142→1694 lines
Move _rush_preselect_and_buy_via_xml, _run_cold_validation_pipeline,
_cold_pipeline_finish_confirm, and _run_warm_validation_pipeline into
FastPipeline. Move get_visible_price_options, _get_visible_price_options_from_xml,
_ocr_price_text_from_card, and _normalize_ocr_price_text into PriceSelector.
Add probe_current_page u2 delegation to PageProbe with Appium fallback.
Add lazy _ensure_pipeline() for test compatibility. Clean up unused imports.
